Create "Toast"

+++ Angular/src/app/date/date-editor/date-editor.component.ts
@@ -6,7 +6,6 @@ import {VisitDate} from "../../backend_types/visit-date";
 import {NavbarService} from "../../service/navbar.service";
 import {NgbDate} from "@ng-bootstrap/ng-bootstrap/datepicker/ngb-date";
 import {HttpErrorResponse} from "@angular/common/http";
-import {MassageResponse} from "../../backend_types/massage-response";
 
 const I18N_VALUES = {
   "ua": {
@@ -129,9 +128,11 @@ export class DateEditorComponent implements OnInit {
       this.loading_save = false;
     }).catch((err: HttpErrorResponse) => {
       this.loading_save = false;
-      alert(
-        ((<MassageResponse> err.error).exceptionMassage != null ? (<MassageResponse> err.error).exceptionMassage : "") + " \n" +
-        ((<MassageResponse> err.error).validationMassage != null ? (<MassageResponse> err.error).validationMassage : ""));
+      if (err.status === 422) {
+        alert(err.error);
+      } else {
+        alert(err.error);
+      }
     });
   }
 
@@ -148,18 +149,28 @@ export class DateEditorComponent implements OnInit {
         ids.push(this.visitDates[i].visitDateId);
       }
     }
-    this.dateService.removeVisitDates(ids).toPromise().then((success: boolean) => {
-      if (success) {
-        alert("Дати успішно видалені!");
-        this.getDates();
+    if (ids.length) {
+      this.dateService.removeVisitDates(ids).toPromise().then((success: boolean) => {
+        if (success) {
+          alert("Дати успішно видалені!");
+          this.getDates();
+          this.loading_del = false;
+        }
+      }).catch((err: HttpErrorResponse) => {
         this.loading_del = false;
-      }
-    }).catch((err: HttpErrorResponse) => {
+        if (err.status === 400) {
+          alert("Видалення неможливе !" + "\n" + err.error);
+        } else {
+          alert(err.error);
+        }
+      });
+    } else if (this.selectedDates.length) {
       this.loading_del = false;
-      alert(
-        ((<MassageResponse> err.error).exceptionMassage != null ? (<MassageResponse> err.error).exceptionMassage : "") + " \n" +
-        ((<MassageResponse> err.error).validationMassage != null ? (<MassageResponse> err.error).validationMassage : ""));
-    });
+      this.getDates();
+    } else {
+      this.getDates();
+      alert("Виберіть хоча б один запис!");
+    }
   }
 
   onCancel() {
